Moved LineChartCard with the rest of the charts
