+++ test/visits/VisitsStats.test.tsx
@@ -4,7 +4,7 @@ import { Mock } from 'ts-mockery';
 import VisitStats from '../../src/visits/VisitsStats';
 import Message from '../../src/utils/Message';
 import { Visit, VisitsInfo } from '../../src/visits/types';
-import LineChartCard from '../../src/visits/helpers/LineChartCard';
+import LineChartCard from '../../src/visits/charts/LineChartCard';
 import VisitsTable from '../../src/visits/VisitsTable';
 import { Result } from '../../src/utils/Result';
 import { Settings } from '../../src/settings/reducers/settings';
