+++ test/visits/charts/LineChartCard.test.tsx
@@ -3,7 +3,7 @@ import { CardHeader, DropdownItem } from 'reactstrap';
 import { Line } from 'react-chartjs-2';
 import { formatISO, subDays, subMonths, subYears } from 'date-fns';
 import { Mock } from 'ts-mockery';
-import LineChartCard from '../../../src/visits/helpers/LineChartCard';
+import LineChartCard from '../../../src/visits/charts/LineChartCard';
 import ToggleSwitch from '../../../src/utils/ToggleSwitch';
 import { NormalizedVisit } from '../../../src/visits/types';
 import { prettify } from '../../../src/utils/helpers/numbers';
